backend/requirements.txt
# Core Framework
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0

# Database Drivers
pymongo==4.6.1

# Elasticsearch
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0

# Redis & Celery
redis==5.0.1
celery==5.3.4
django-redis==5.4.0

# WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
python-jose==3.3.0
cryptography==41.0.7

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
isort==5.13.2

# WSGI Server
gunicorn==21.2.0
whitenoise==6.6.0
